Move functions/src/supabase to backend/shared/supabase
