+++ backend/functions/src/score-contracts.ts
@@ -6,8 +6,8 @@ import { Contract } from 'common/contract'
 import { loadPaginated, log } from 'shared/utils'
 import { removeUndefinedProps } from 'common/util/object'
 import { DAY_MS, HOUR_MS } from 'common/util/time'
-import { createSupabaseClient } from './supabase/init'
-import { getRecentContractLikes } from './supabase/likes'
+import { createSupabaseClient } from 'shared/supabase/init'
+import { getRecentContractLikes } from 'shared/supabase/likes'
 import { run } from 'common/supabase/utils'
 import { logit } from 'common/util/math'
