+++ backend/functions/src/scripts/test-supabase.ts
@@ -1,9 +1,9 @@
-import { initAdmin } from '../scripts/script-init'
+import { initAdmin } from './script-init'
 initAdmin()
 
 import { DAY_MS } from 'common/util/time'
-import { createSupabaseClient } from '../supabase/init'
-import { getRecentContractLikes } from '../supabase/likes'
+import { createSupabaseClient } from 'shared/supabase/init'
+import { getRecentContractLikes } from 'shared/supabase/likes'
 
 const main = async () => {
   const db = createSupabaseClient()
