+++ backend/functions/src/update-recommended.ts
@@ -8,7 +8,7 @@ import { newEndpointNoAuth } from './api'
 import { getMarketRecommendations } from 'common/recommendation'
 import { run } from 'common/supabase/utils'
 import { mapAsync } from 'common/util/promise'
-import { createSupabaseClient } from './supabase/init'
+import { createSupabaseClient } from 'shared/supabase/init'
 import { buildArray, filterDefined } from 'common/util/array'
 import { Contract } from 'common/contract'
 import { chooseRandomSubset } from 'common/util/random'
