+++ backend/functions/src/weekly-portfolio-updates.ts
@@ -6,7 +6,7 @@ import { PrivateUser } from 'common/user'
 import { getUsersContractMetrics } from 'common/supabase/contract-metrics'
 import { createWeeklyPortfolioUpdateNotification } from './create-notification'
 import { getUsernameById } from 'common/supabase/users'
-import { createSupabaseClient } from './supabase/init'
+import { createSupabaseClient } from 'shared/supabase/init'
 import { log } from 'shared/utils'
 import { WeeklyPortfolioUpdate } from 'common/weekly-portfolio-update'
 import { DAY_MS } from 'common/util/time'
